requirements.txt
flask
flask_sqlalchemy
gunicorn==20.1.0
python-dotenv
psycopg2-binary
